services logic added

+++ src/main/java/com/xyz/mobilehandsetservice/dto/Release.java
@@ -1,11 +1,9 @@
 package com.xyz.mobilehandsetservice.dto;
 
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 @Data
 public class Release {
     public String announceDate;
-    public BigDecimal priceEur;
+    public Double priceEur;
 }
