+++ src/main/java/com/xyz/mobilehandsetservice/entity/ReleaseVO.java
@@ -1,11 +1,9 @@
 package com.xyz.mobilehandsetservice.entity;
 
-import java.math.BigDecimal;
-
 import lombok.Data;
 
 @Data
 public class ReleaseVO {
     public String announceDate;
-    public BigDecimal priceEur;
+    public Double priceEur;
 }
